requirements.txt
aiohttp==3.8.4
babel==2.12.1
beautifulsoup4==4.12.0
branca==0.6.0
celery==5.2.7
configparser==5.3.0
django-ckeditor==6.5.1
django-debug-toolbar==3.8.1
django-multiupload==0.6.1
django-scheduler==0.10.1
docopt==0.6.2
flask-markdown==0.3
frozen-flask==0.18
gevent==22.10.2
habanero==1.2.3
hitran-api==1.2.2.0
hjson==3.1.0
html5lib==1.1
httptools==0.5.0
humanize==4.6.0
icecream==2.1.3
ipympl==0.9.3
ipyvue==1.9.0
ipywebrtc==0.6.0
joblib==1.2.0
json-tricks==3.16.1
keyring==23.13.1
lmfit==1.1.0
lxml==4.9.2
mpldatacursor==0.7.1
peakutils==1.3.4
pika==1.3.1
pip-chill==1.0.1
plotly==5.13.1
psycopg-binary==3.1.8
psycopg2-binary==2.9.5
publib==0.3.3
pyrogram==2.0.102
pytest==7.2.2
python-dotenv==0.21.1
pyvo==1.4.1
redis==4.5.1
schedule==1.1.0
seaborn==0.12.2
signals==0.0.2
tables==3.8.0
termcolor==2.2.0
tgcrypto==1.2.5
traittypes==0.2.1
uvloop==0.17.0
vaex-astro==0.9.3
vaex-hdf5==0.14.1
vaex-ml==0.18.1
vaex-server==0.8.1
vaex-viz==0.5.4
watchfiles==0.18.1
websockets==10.4
xarray==2023.1.0
xyzservices==2023.2.0
